+++ packages/jsapi-components/src/useViewportData.ts
@@ -13,11 +13,12 @@ import {
   useOnScrollOffsetChangeCallback,
   WindowedListData,
 } from '@deephaven/react-hooks';
-import { KeyedItem, SCROLL_DEBOUNCE_MS } from '@deephaven/utils';
+import { KeyedItem } from '@deephaven/utils';
 import useInitializeViewportData from './useInitializeViewportData';
 import useSetPaddedViewportCallback from './useSetPaddedViewportCallback';
 import useTableSize from './useTableSize';
 import useTableListener from './useTableListener';
+import { SCROLL_DEBOUNCE_MS } from './Constants';
 
 const log = Log.module('useViewportData');